services/embedding_service/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.6.0
python-dotenv>=1.0.0
httpx>=0.26.0
pymongo>=4.6.0
sentence-transformers>=2.3.0
numpy>=1.26.0
motor>=3.4.0
torch>=2.1.0
transformers>=4.36.0
